feat: adds update drug route
fix: fixes improper error handling in drug handler

+++ handlers/drug_handler.go
@@ -26,8 +26,9 @@ func CreateDrug(h *DrugHandler) gin.HandlerFunc {
 
 		if result.Error != nil {
 			ctx.JSON(http.StatusInternalServerError, gin.H{
-				"error": result.Error,
+				"error": result.Error.Error(),
 			})
+			return
 		}
 
 		ctx.JSON(http.StatusOK, gin.H{
@@ -44,8 +45,9 @@ func GetDrugs(h *DrugHandler) gin.HandlerFunc {
 
 		if result.Error != nil {
 			ctx.JSON(http.StatusInternalServerError, gin.H{
-				"error": result.Error,
+				"error": result.Error.Error(),
 			})
+			return
 		}
 
 		ctx.JSON(http.StatusOK, gin.H{
@@ -65,18 +67,69 @@ func DeleteDrug(h *DrugHandler) gin.HandlerFunc {
 			ctx.JSON(http.StatusInternalServerError, gin.H{
 				"error": result.Error,
 			})
+			return
 		}
 
-		if drug.ID == 0 {
+		if result.Error == gorm.ErrRecordNotFound {
 			ctx.JSON(http.StatusInternalServerError, gin.H{
 				"error": "Drug not found",
 			})
+			return
 		}
 
 		h.DB.Delete(&drug)
 
 		ctx.JSON(http.StatusOK, gin.H{
 			"message": "Drug deleted succesfully",
+		})
+	}
+}
+
+func UpdateDrug(h *DrugHandler) gin.HandlerFunc {
+	return func(ctx *gin.Context) {
+		drugId := ctx.Param("id")
+
+		var updatedDrug models.Drug
+		if err := ctx.ShouldBindJSON(&updatedDrug); err != nil {
+			ctx.JSON(http.StatusBadRequest, gin.H{
+				"error": "Invalid request body",
+			})
+			return
+		}
+
+		var existingDrug models.Drug
+
+		result := h.DB.First(&existingDrug, drugId)
+		if result.Error != nil {
+			ctx.JSON(http.StatusInternalServerError, gin.H{
+				"error": result.Error.Error(),
+			})
+			return
+		}
+
+		if result.Error == gorm.ErrRecordNotFound {
+			ctx.JSON(http.StatusInternalServerError, gin.H{
+				"error": "Drug not found",
+			})
+			return
+		}
+
+		existingDrug.Name = updatedDrug.Name
+		existingDrug.Approved = updatedDrug.Approved
+		existingDrug.MinDose = updatedDrug.MinDose
+		existingDrug.MaxDose = updatedDrug.MaxDose
+		existingDrug.AvailableAt = updatedDrug.AvailableAt
+
+		saveResult := h.DB.Save(&existingDrug)
+		if saveResult.Error != nil {
+			ctx.JSON(http.StatusInternalServerError, gin.H{
+				"error": saveResult.Error.Error(),
+			})
+			return
+		}
+
+		ctx.JSON(http.StatusOK, gin.H{
+			"drug": existingDrug,
 		})
 	}
 }
